Script to report on team/member folders
Example script to iterate over entire team or a specific subset specified via a CSV file. 
Per user it will report on all folders, the sum of data in each folder & its sub folders, sum of data in folder excluding sub-folders. Number of files and folders in each folder. 

Depending on size of team this could be a very slow process! So targeting a specific user or small subset of users.

There's no threading in this example.